Move DownloadButton into it's own file
